Termo do transporte de especies add na equacao de energia. Erro a funcao advecNdim corrigido.

+++ include/StructDefine.h
@@ -4,7 +4,7 @@
   #include<HccaStdBool.h>
   #include<Define.h>
 
-  DOUBLE gStep;
+  INT gStep;
 
 /*...*/
   typedef struct{
